rename: #38 components 디렉토리로 모든 컴포넌트 클래스 이동
