+++ AutoBid_FE/main.css
@@ -1,2 +1,2 @@
-@import "./src/app.css";
-@import "./src/Counter/counter.css";
+@import "./src/component/app.css";
+@import "./src/component/Counter/counter.css";
