+++ AutoBid_FE/main.ts
@@ -1,4 +1,4 @@
-import App from "./src/App";
+import App from "./src/component/App";
 
 const $app = document.getElementById('App') as HTMLElement;
 new App($app, {});
