+++ AutoBid_FE/src/component/Counter/counter.css
